Move Scratchpad provider to same directory as others
