+++ graylog2-web-interface/src/components/common/Scratchpad.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'wrappedTestingLibrary';
 
-import { ScratchpadProvider } from 'providers/ScratchpadProvider';
+import { ScratchpadProvider } from 'contexts/ScratchpadProvider';
 
 import Scratchpad from './Scratchpad';
