+++ graylog2-web-interface/src/components/navigation/ScratchpadToggle.jsx
@@ -1,9 +1,9 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
 
+import { ScratchpadContext } from 'contexts/ScratchpadProvider';
 import { Button } from 'components/graylog';
 import { Icon } from 'components/common';
-import { ScratchpadContext } from 'providers/ScratchpadProvider';
 
 const Toggle = styled(Button)`
   padding-left: 6px;
