+++ graylog2-web-interface/src/contexts/ScratchpadProvider.jsx
