+++ graylog2-web-interface/src/routing/App.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import styled, { css } from 'styled-components';
 import chroma from 'chroma-js';
 
+import { ScratchpadProvider } from 'contexts/ScratchpadProvider';
 import { Scratchpad, Icon, Spinner } from 'components/common';
-import { ScratchpadProvider } from 'providers/ScratchpadProvider';
 import CurrentUserContext from 'contexts/CurrentUserContext';
 import Navigation from 'components/navigation/Navigation';
 import ReportedErrorBoundary from 'components/errors/ReportedErrorBoundary';
